+++ src/api/storage/IStorage.ts
@@ -1,4 +1,4 @@
-import { IApiReferenceIndex, IApiReferenceLibrary } from "./IApiReference";
+import { IApiReferenceIndex, IApiReferenceLibrary } from "../IApiReference";
 
 /**
  * Api modules storage
@@ -7,4 +7,6 @@ export default abstract class IStorage {
 	abstract getApiIndex(): Promise<IApiReferenceIndex>;
 
 	abstract getLibrary(id: string): Promise<IApiReferenceLibrary>;
+
+	abstract clear(): Promise<any>;
 }