+++ CHANGELOG.md
@@ -1,6 +1,27 @@
 # Change Log
 All notable changes to the **UI5 Explorer** extension will be documented here.
 
+## 0.4.0 - 2019-12-20
+- Added support for UI5 >= 1.63
+- Undefined description is not rendered anymore
+- Updated axios version
+
+## 0.3.0 - 2018-07-29
+- Configurable UI5 Version (currently supported only for 1.54.* versions due to different file structure for older versions)
+- Reload tree after configuration change
+- Added file cache
+- New **Clear Cache** command
+- Bugfixes (eg. for *sap.ui.base.EventProvider*)
+
+## 0.2.0 - 2018-07-04
+- Collapsible/Expandable methods and events
+- Quick navigation to sections under the header
+- "Back to top" button at the bottom right corner
+- Better links support (aggregations, associations, method/event parameters...)
+- Added support for methods without parameters
+- Added "default" aggragation marker
+- Several bug fixes
+
 ## 0.1.0 - 2018-06-27
 - Marketplace adjustment (icon, description, links...)